pkg/kubelet/userns: Provide stub implementation for windows
