+++ pkg/kubelet/userns/userns_manager_switch_test.go
@@ -1,3 +1,6 @@
+//go:build !windows
+// +build !windows
+
 /*
 Copyright 2024 The Kubernetes Authors.
